requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.5.2               # via kombu
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
billiard==3.6.3.0         # via celery
brotlipy==0.7.0           # via -r requirements.in
celery[redis]==4.4.2      # via -r requirements.in
certifi==2019.11.28       # via sentry-sdk
cffi==1.14.0              # via brotlipy
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via odfpy
diff-match-patch==20181111  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-import-export==2.0.2  # via -r requirements.in
django-js-reverse==0.9.1  # via -r requirements.in
django-log-request-id==1.4.1  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-webpack-loader==0.7.0  # via -r requirements.in
django==2.2.11            # via -r requirements.in, django-import-export, django-js-reverse, django-log-request-id, django-model-utils
et-xmlfile==1.0.1         # via openpyxl
gunicorn==20.0.4          # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.13.0           # via -r requirements.in
jdcal==1.4.1              # via openpyxl
jedi==0.16.0              # via ipython
kombu==4.6.8              # via celery
markuppy==1.14            # via tablib
odfpy==1.4.1              # via tablib
openpyxl==3.0.3           # via tablib
parso==0.6.2              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==3.0.4     # via ipython
psutil==5.7.0             # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pycparser==2.20           # via cffi
pygments==2.6.1           # via ipython
python-decouple==3.3      # via -r requirements.in
pytz==2019.3              # via celery, django
pyyaml==5.3.1             # via tablib
redis==3.4.1              # via celery
sentry-sdk==0.14.3        # via -r requirements.in
six==1.14.0               # via traitlets
sqlparse==0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]==1.1.0  # via django-import-export
traitlets==4.3.3          # via ipython
urllib3==1.25.8           # via sentry-sdk
vine==1.3.0               # via amqp, celery
wcwidth==0.1.9            # via prompt-toolkit
whitenoise==5.0.1         # via -r requirements.in
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
